+++ src/Popups/Credits.js
@@ -13,6 +13,7 @@ import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import Alert from '@material-ui/lab/Alert';
 import { makeStyles } from '@material-ui/core/styles';
+import log from '../util/logger.js';
 
 
 const useStyles = makeStyles(theme => ({
@@ -57,6 +58,15 @@ function CreditsPopup(props) {
       handleClose();
       props.openTutorial();
     }
+    else if (newValue === 3) {
+      const data = JSON.stringify({logs: log.export(), localStorage: localStorage});
+      const blob = new Blob([data], {type: 'text/json'});
+      const a = document.createElement('a');
+      a.download = 'debug.json';
+      a.href = window.URL.createObjectURL(blob)
+      a.dataset.downloadurl =  ['text/json', a.download, a.href].join(':');
+      a.click();
+    }
     else {
       setExpanded(newValue);
     }
@@ -82,6 +92,7 @@ function CreditsPopup(props) {
           <Tab label="Changelog" />
           <Tab label="Credits" />
           <Tab label="Tutorial" />
+          <Tab label="Debug" />
         </Tabs>
         <IconButton className={classes.closeButton} onClick={handleClose}>
           <CloseIcon />
@@ -89,6 +100,23 @@ function CreditsPopup(props) {
       </DialogTitle>
       <DialogContent dividers className={classes.dialog}>
         <div hidden={expanded !== 0}>
+          <Paper className={classes.content}>
+            <Typography variant="h5" className={classes.version}>v1.4.1 (2021-04-16)</Typography>
+            <Typography variant="h6">Added</Typography>
+            <List dense>
+              <ListItem>Debug button: allow any user to easily export debug information, to help investigating bugs. The new button is accessible via the changelog & credits popup</ListItem>
+            </List>
+            <Typography variant="h6">Changed</Typography>
+            <List dense>
+              <ListItem>Behind the scenes: removed proxy, thanks to a welcomed change on FSE side regarding CORS headers</ListItem>
+            </List>
+            <Typography variant="h6">Fixed</Typography>
+            <List dense>
+              <ListItem>Bug in "From ICAO" and "To ICAO" filters, that would wrongly hide some jobs</ListItem>
+              <ListItem>Bug in Route Finder, that would prevent the search from finishing</ListItem>
+            </List>
+          </Paper>
+
           <Paper className={classes.content}>
             <Typography variant="h5" className={classes.version}>v1.4.0 (2021-03-13)</Typography>
             <Typography variant="h6">Added</Typography>